requirements.txt
certifi==2017.11.5
chardet==3.0.4
codecov==2.0.9
coverage==4.4.2
Django==1.11.7
django-configurations==2.0
django-constance==2.0.0
django-jenkins==0.110.0
django-picklefield==1.0.0
djangorestframework==3.7.3
idna==2.6
parameterized==0.6.1
pytz==2017.3
requests==2.18.4
urllib3==1.22
